+++ src/containers/WebComponentLoader.jsx
@@ -1,24 +1,64 @@
-import React, { useEffect } from "react";
+import React, { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
-import { setProject, setSenseHatAlwaysEnabled } from "../redux/EditorSlice";
+import { setSenseHatAlwaysEnabled } from "../redux/EditorSlice";
 import WebComponentProject from "../components/WebComponentProject/WebComponentProject";
 import { useTranslation } from "react-i18next";
 import { setInstructions } from "../redux/InstructionsSlice";
+import { useProject } from "../hooks/useProject";
+import { useProjectPersistence } from "../hooks/useProjectPersistence";
+import { removeUser, setUser } from "../redux/WebComponentAuthSlice";
 
 const WebComponentLoader = (props) => {
   const loading = useSelector((state) => state.editor.loading);
-  const { code, senseHatAlwaysEnabled = false, instructions } = props;
+  const {
+    authKey,
+    identifier,
+    code,
+    senseHatAlwaysEnabled = false,
+    instructions,
+  } = props;
   const dispatch = useDispatch();
   const { t } = useTranslation();
+  const [projectIdentifier, setProjectIdentifier] = useState(identifier);
+  const project = useSelector((state) => state.editor.project);
+  const user = JSON.parse(localStorage.getItem(authKey));
+  const justLoaded = useSelector((state) => state.editor.justLoaded);
+  const hasShownSavePrompt = useSelector(
+    (state) => state.editor.hasShownSavePrompt,
+  );
+  const saveTriggered = useSelector((state) => state.editor.saveTriggered);
 
   useEffect(() => {
-    const proj = {
-      type: "python",
-      components: [{ name: "main", extension: "py", content: code }],
-    };
+    if (user) {
+      dispatch(setUser(user));
+    } else {
+      dispatch(removeUser());
+    }
+  }, [user, dispatch]);
+
+  useEffect(() => {
+    if (loading === "idle" && project.identifier) {
+      setProjectIdentifier(project.identifier);
+    }
+  }, [loading, project]);
+
+  useProject({
+    projectIdentifier: projectIdentifier,
+    code,
+    accessToken: user && user.access_token,
+  });
+
+  useProjectPersistence({
+    user,
+    project,
+    justLoaded,
+    hasShownSavePrompt,
+    saveTriggered,
+  });
+
+  useEffect(() => {
     dispatch(setSenseHatAlwaysEnabled(senseHatAlwaysEnabled));
-    dispatch(setProject(proj));
-  }, [code, senseHatAlwaysEnabled, dispatch]);
+  }, [senseHatAlwaysEnabled, dispatch]);
 
   useEffect(() => {
     if (instructions) {