+++ CHANGELOG.md
@@ -11,10 +11,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Move local development port to avoid projects-ui clash (#736)
 - Update design-system-react dependency to publicly available version
 
+### Fixed
+
+- Local docker setup for development (#739)
+
 ## [0.19.4] - 2023-11-08
 
 ### Changed
 
+- Added p5 library that can access WebComponent shadowRoot (#731)
 - Allow external rpf.io links (#729)
 
 ## [0.19.3] - 2023-10-25
@@ -25,6 +30,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Web component tests (#709, #710)
 - `instructions` attribute for the web component (#712)
 - Instructions slice to store data passed from the Projects site (#712)
+- Adding auth to the web component (#728)
+- Allow web component to load, save and remix projects (#728)
 
 ### Changed